refactor(#14): rename api file
