+++ site/api/requests/getPaginationInfo.ts
@@ -1,4 +1,4 @@
-import { buildApiRequest } from '../api';
+import { buildApiRequest } from '../utils';
 import {
 	GetPaginationInfoRequest,
 	GetPaginationInfoResponse,
