+++ site/api/requests/getPost.ts
@@ -1,4 +1,4 @@
-import { buildApiRequest } from '../api';
+import { buildApiRequest } from '../utils';
 import { Post } from '../../../common/Post';
 import { GetPostRequest } from '../../../common/api/GetPost';
