+++ site/api/requests/getPostUrls.ts
@@ -1,3 +1,3 @@
-import { buildApiRequest } from '../api';
+import { buildApiRequest } from '../utils';
 
 export const getPostUrls = buildApiRequest<void, string[]>('/getPostUrls');
